% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_1_data_function.R
\name{Q_1_data}
\alias{Q_1_data}
\title{Evaluate the critical point and observed value of the test statistic Q_1 from a given data}
\usage{
Q_1_data(data, a, b, size)
}
\arguments{
\item{data}{a real data set for a Two-way additive ANOVA model}

\item{a}{number of levels of the row factor A}

\item{b}{number of levels of the column factor B}

\item{size}{the level of significance}
}
\value{
numeric vector consisting of observed and critical values of Q_1
}
\description{
Evaluate the critical point and observed value of the test statistic Q_1 from a given data
}
\examples{
data<-read.csv("https://raw.githubusercontent.com/RajuProbStat/AdditiveANOVA/main/Alcohol_real_data.csv")
Q_1_data(data,3,2,0.05)
}
