% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_2_3_function.R
\name{Q_2_3}
\alias{Q_2_3}
\title{Evaluate the critical point and observed value of the test statistic Q_2 and Q_3}
\usage{
Q_2_3(a, b, size, n, mean, var)
}
\arguments{
\item{a}{number of levels of the row factor A}

\item{b}{number of levels of the column factor}

\item{size}{the level of significance}

\item{n}{matrix having a rows and b columns, and (i,j)-th element is the sample size in the (i,j)-th cell}

\item{mean}{matrix having a rows and b columns, and (i,j)-th element is the sample mean in the (i,j)-th cell}

\item{var}{matrix having a rows and b columns, and (i,j)-th element is the sample variance in the (i,j)-th cell}
}
\value{
numeric vector consisting of observed and critical values of Q_2 and Q_3
}
\description{
Evaluate the critical point and observed value of the test statistic Q_2 and Q_3
}
\examples{
Q_2_3(3,2,0.05,rbind(c(15,17),c(16,14),c(18,15)),rbind(c(186.67,332.24),c(261.12,485.50),c(302.44,581.87)),rbind(c(295.10,2374.94),c(372.92,1008.27),c(840.85,661.12)))
}
