% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_2_3_data_function.R
\name{Q_2_3_data}
\alias{Q_2_3_data}
\title{Evaluate the critical points and observed values of the test statistics Q_2 and Q_3 from a given data}
\usage{
Q_2_3_data(data, a, b, size)
}
\arguments{
\item{data}{a real data set for a Two-way additive ANOVA model}

\item{a}{number of levels of the row factor A}

\item{b}{number of levels of the column factor B}

\item{size}{the level of significance}
}
\value{
numeric vector consisting of observed and critical values of Q_2 and Q_3
}
\description{
Evaluate the critical points and observed values of the test statistics Q_2 and Q_3 from a given data
}
\examples{
data<-read.csv("https://raw.githubusercontent.com/RajuProbStat/AdditiveANOVA/main/Alcohol_real_data.csv")
Q_2_3_data(data,3,2,0.05)
}
